# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Autonomous YouTube Factory API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List asset results
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
        - name: storyboardId
          in: query
          schema:
            type: ["string", "null"]
        - name: sceneId
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: assetKind
          in: query
          schema:
            type: ["string", "null"]
        - name: provider
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Asset list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetList"
    post:
      operationId: acquireAssets
      tags: [assets]
      summary: Enqueue asset acquisition for a storyboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "202":
          description: Asset acquisition queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetAcquireResponse"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get asset result by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Asset result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetResult"
        "404":
          description: Not found
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete an asset result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /renders:
    get:
      operationId: listRenders
      tags: [renders]
      summary: List render jobs (the Render Queue)
      parameters:
        - name: timelineId
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Render list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenderList"
    post:
      operationId: startRender
      tags: [renders]
      summary: Enqueue a render job for a completed timeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenderInput"
      responses:
        "202":
          description: Render accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenderResult"
  /renders/{id}:
    get:
      operationId: getRender
      tags: [renders]
      summary: Get a render result by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Render result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenderResult"
        "404":
          description: Not found
    delete:
      operationId: deleteRender
      tags: [renders]
      summary: Delete a render record (and its output file, if present)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectList"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /projects/{id}/run:
    post:
      operationId: runProject
      tags: [projects]
      summary: Start the full pipeline for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Pipeline started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineRun"

  # ── Pipelines ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, title, status, topic, createdAt, updatedAt]
      properties:
        id:          { type: string }
        title:       { type: string }
        topic:       { type: string }
        description: { type: ["string", "null"] }
        status:
          type: string
          enum: [draft, queued, running, completed, failed, cancelled]
        thumbnailUrl:  { type: ["string", "null"] }
        videoUrl:      { type: ["string", "null"] }
        youtubeId:     { type: ["string", "null"] }
        youtubeUrl:    { type: ["string", "null"] }
        pipelineId:    { type: ["string", "null"] }
        tags:
          type: array
          items: { type: string }
        createdAt:   { type: string }
        updatedAt:   { type: string }
    ProjectInput:
      type: object
      required: [title, topic]
      properties:
        title:       { type: string }
        topic:       { type: string }
        description: { type: string }
        tags:
          type: array
          items: { type: string }
    ProjectUpdate:
      type: object
      properties:
        title:       { type: string }
        topic:       { type: string }
        description: { type: string }
        status:      { type: string }
        tags:
          type: array
          items: { type: string }
    ProjectList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Project" }
        total: { type: integer }

    # ── Pipeline ─────────────────────────────────────────────────────────────
    PipelineRun:
      type: object
      required: [id, projectId, status, stages, startedAt]
      properties:
        id:          { type: string }
        projectId:   { type: string }
        status:
          type: string
          enum: [queued, running, completed, failed, cancelled]
        currentStage: { type: ["string", "null"] }
        progress:     { type: integer }
        stages:
          type: array
          items: { $ref: "#/components/schemas/PipelineStage" }
        startedAt:    { type: string }
        completedAt:  { type: ["string", "null"] }
        errorMessage: { type: ["string", "null"] }
    PipelineStage:
      type: object
      required: [name, status, order]
      properties:
        name:       { type: string }
        status:
          type: string
          enum: [pending, running, completed, failed, skipped]
        order:      { type: integer }
        startedAt:  { type: ["string", "null"] }
        completedAt: { type: ["string", "null"] }
        durationMs: { type: ["integer", "null"] }
        error:      { type: ["string", "null"] }
...
    AssetInput:
      type: object
      required: [storyboardId]
      properties:
        storyboardId:  { type: string }
        sceneIds:
          type: array
          items: { type: string }
          description: Empty = all scenes in the storyboard
        assetKinds:
          type: array
          items:
            type: string
            enum: [image, video, svg, chart, map, icon]
        providers:
          type: array
          items: { type: string }
          description: Ordered preference list — stock providers searched first, then generation
        forceGenerate: { type: boolean }
    AssetResult:
      type: object
      required: [id, storyboardId, sceneId, assetKind, status, license, tags, logs, createdAt, updatedAt]
      properties:
        id:                  { type: string }
        storyboardId:        { type: string }
        sceneId:             { type: string }
        assetKind:
          type: string
          enum: [image, video, svg, chart, map, icon]
        provider:            { type: ["string", "null"] }
        providerAssetId:     { type: ["string", "null"] }
        sourceUrl:           { type: ["string", "null"] }
        license:
          type: string
          enum: [public_domain, cc0, cc_by, cc_by_sa, commercial, generated, mit, unknown]
        prompt:              { type: ["string", "null"] }
        negativePrompt:      { type: ["string", "null"] }
        generationParameters:
          type: object
          additionalProperties: {}
        width:               { type: ["integer", "null"] }
        height:              { type: ["integer", "null"] }
        aspectRatio:         { type: ["string", "null"] }
        status:
          type: string
          enum: [pending, searching, downloading, generating, ready, failed, cached]
        costEstimateUsd:     { type: ["number", "null"] }
        generationTimeMs:    { type: ["integer", "null"] }
        fileSizeBytes:       { type: ["integer", "null"] }
        localCachePath:      { type: ["string", "null"] }
        thumbnailPath:       { type: ["string", "null"] }
        tags:
          type: array
          items: { type: string }
        qualityScore:        { type: ["number", "null"] }
        relevanceScore:      { type: ["number", "null"] }
        logs:
          type: array
          items: { type: string }
        errorMessage:        { type: ["string", "null"] }
        createdAt:           { type: string }
        updatedAt:           { type: string }
        completedAt:         { type: ["string", "null"] }
    AssetList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/AssetResult" }
        total: { type: integer }
    AssetAcquireResponse:
      type: object
      required: [items, total, message]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/AssetResult" }
        total:   { type: integer }
        message: { type: string }
...
    RenderInput:
      type: object
      required: [timelineId]
      properties:
        timelineId:  { type: string }
        voiceId:     { type: ["string", "null"] }
        resolution:
          type: string
          enum: ["720p", "1080p", "4k"]
        fps:         { type: integer, minimum: 24, maximum: 60 }
        aspectRatio:
          type: string
          enum: ["16:9", "9:16", "1:1"]
        cropMode:
          type: string
          enum: [safe_crop, letterbox, blur_pad]
        hardwareAcceleration: { type: boolean }
        addBackgroundMusic:   { type: boolean }
        musicVolume:          { type: number, minimum: 0, maximum: 1 }
        generatePreview:      { type: boolean }
    RenderOutput:
      type: object
      properties:
        localPath:        { type: ["string", "null"] }
        fileSizeBytes:    { type: integer }
        durationSeconds:  { type: number }
        width:            { type: integer }
        height:           { type: integer }
        fps:              { type: integer }
        codec:            { type: string }
        audioCodec:       { type: string }
        format:           { type: string }
    RenderStats:
      type: object
      properties:
        renderTimeSeconds: { type: number }
        framesEncoded:     { type: integer }
        encodeFps:         { type: number }
        realtimeFactor:    { type: number }
        retries:           { type: integer }
    RenderMetadata:
      type: object
      properties:
        sceneCount:            { type: integer }
        clipCount:             { type: integer }
        placeholderClipCount:  { type: integer }
        hasNarration:          { type: boolean }
        hasBackgroundMusic:    { type: boolean }
        sourceTimelineId:      { type: string }
        sourceVoiceId:         { type: ["string", "null"] }
    RenderResult:
      type: object
      required: [id, timelineId, status, logs, createdAt, updatedAt]
      properties:
        id:              { type: string }
        timelineId:      { type: string }
        voiceId:         { type: ["string", "null"] }
        status:
          type: string
          enum: [pending, running, completed, failed]
        progress:        { type: integer }
        resolution:      { type: string }
        width:           { type: integer }
        height:          { type: integer }
        fps:             { type: integer }
        aspectRatio:     { type: string }
        cropMode:        { type: string }
        hardwareAcceleration: { type: boolean }
        renderPlan:      { type: object, additionalProperties: {} }
        renderOutput:    { $ref: "#/components/schemas/RenderOutput" }
        renderStats:     { $ref: "#/components/schemas/RenderStats" }
        renderMetadata:  { $ref: "#/components/schemas/RenderMetadata" }
        previewOutput:   { $ref: "#/components/schemas/RenderOutput" }
        logs:
          type: array
          items: { type: string }
        errorMessage:    { type: ["string", "null"] }
        createdAt:       { type: string }
        updatedAt:       { type: string }
        completedAt:     { type: ["string", "null"] }
    RenderList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/RenderResult" }
        total: { type: integer }
...